Add member_stats as Cog and remove redundant lines
