+++ bot.py
@@ -7,9 +7,8 @@ from cogs.send_welcome import Welcome
 from cogs.validation import ModeratorChecking
 from cogs.error_handler import CommandErrorHandler
 from cogs.help import GetHelp
+from cogs.member_stats import MemberStats
 # from .validation import is_moderator, is_mod_commands_channel
-
-# Add in intent to allow us to monitor member stats
 
 intents = discord.Intents.default()
 intents.members = True
@@ -19,41 +18,6 @@ bot = commands.Bot(command_prefix='$')
 @bot.event
 async def on_ready():
     print("Ready to go!")
-
-class MemberStats(commands.Cog):
-    def __init__(self, bot):
-        self.bot = bot
-
-    @commands.command()
-
-    async def mstats(self, ctx):
-
-        #if not await is_moderator(ctx):
-        #    return
-        #if not await is_mod_commands_channel(ctx):
-        #    return
-
-        with open('./server_specific/channel_ids.json', 'r') as id_file:
-            channel_id_dict = json.loads(id_file.read())
-            
-            guild_id = channel_id_dict['GUILD']
-            welcome_id = channel_id_dict['WELCOME']
-            mbc_id = channel_id_dict['MOD_COMMANDS']
-
-            guild = self.bot.get_guild(guild_id)
-            welcome_channel = guild.get_channel(welcome_id)
-     
-            true_member_count = 0
-            member_online_count = 0
-
-            async for member in guild.fetch_members(limit=None):
-                if not member.bot:
-                    true_member_count += 1
-                    if member.status != "offline":
-                        member_online_count += 1
-
-            await ctx.send(f'Total members: {true_member_count}')
-            await ctx.send(f'Online members: {member_online_count}')
 
 # Read in the token.
 with open(Path('./private/priv_data.json'), 'r') as data_file:
